PErmission changes

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class CategoryController extends Controller
@@ -21,19 +22,19 @@ class CategoryController extends Controller
         // $allcategories = Category::with('parent')->latest()->get();
         // $allcategories = Category::with('parent')->latest()->get()
         $allcategories = Category::with('parent')
-        ->orderBy('created_at', 'desc')
-        ->get() // Get the collection
-        ->map(function ($category) {
-            return [
-                'id' => $category->id,
-                'name' => $category->name,
-                'parent' => $category->parent ? [
-                    'id' => $category->parent->id,
-                    'name' => $category->parent->name,
-                ] : null,
-                'hierarchy_string' => $category->hierarchy_string, // Add this
-            ];
-        });
+            ->orderBy('created_at', 'desc')
+            ->get() // Get the collection
+            ->map(function ($category) {
+                return [
+                    'id' => $category->id,
+                    'name' => $category->name,
+                    'parent' => $category->parent ? [
+                        'id' => $category->parent->id,
+                        'name' => $category->parent->name,
+                    ] : null,
+                    'hierarchy_string' => $category->hierarchy_string, // Add this
+                ];
+            });
 
 
         return Inertia::render('Categories/Index', [
@@ -54,6 +55,10 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'parent_id' => 'nullable|exists:categories,id',
@@ -76,6 +81,9 @@ class CategoryController extends Controller
 
     public function update(Request $request, Category $category)
     {
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'parent_id' => 'nullable|exists:categories,id',
@@ -103,7 +111,9 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
-
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         $category->delete();
         return redirect()->route('categories.index')->banner('Category Deleted successfully.');
     }
