+++ app/Http/Controllers/SupplierController.php
@@ -1,10 +1,12 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Category;
 use App\Models\Product;
 use App\Models\Supplier;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 
@@ -33,8 +35,10 @@ class SupplierController extends Controller
 
     public function store(Request $request)
     {
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
 
- 
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'contact' => 'required|string|max:20',
@@ -57,7 +61,7 @@ class SupplierController extends Controller
             $fileExtension = $request->file('image')->getClientOriginalExtension();
             $fileName = 'supplier_' . date("YmdHis") . '.' . $fileExtension;
             $path = $request->file('image')->storeAs('suppliers', $fileName, 'public');
-            $validated['image'] = 'storage/'.$path;
+            $validated['image'] = 'storage/' . $path;
         }
 
         Supplier::create($validated);
@@ -69,7 +73,9 @@ class SupplierController extends Controller
     public function update(Request $request, Supplier $supplier)
     {
 
-
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
         // Validate incoming data
         $validated = $request->validate([
             'name' => 'nullable|string|max:255',
@@ -78,7 +84,6 @@ class SupplierController extends Controller
             'address' => 'nullable|string|max:500',
             'image' => 'nullable|mimes:jpeg,png,jpg,gif|max:2048',
         ]);
-
 
 
         if ($request->hasFile('image')) {
@@ -98,15 +103,11 @@ class SupplierController extends Controller
         }
 
 
-
         $supplier->update($validated);
 
 
         // Redirect back with success message
         return redirect()->route('suppliers.index')->banner('Supplier updated successfully.');
-
-
-
     }
 
 
@@ -115,13 +116,16 @@ class SupplierController extends Controller
 
     public function destroy(Supplier $supplier)
     {
+        if (!Gate::allows('hasRole', ['Admin'])) {
+            abort(403, 'Unauthorized');
+        }
+
         if ($supplier->image && Storage::disk('public')->exists(str_replace('storage/', '', $supplier->image))) {
             Storage::disk('public')->delete(str_replace('storage/', '', $supplier->image));
         }
 
-         $supplier->delete();
+        $supplier->delete();
 
-         return redirect()->route('suppliers.index')->banner('Supplier deleted successfully.');
-     }
-
+        return redirect()->route('suppliers.index')->banner('Supplier deleted successfully.');
+    }
 }
